Batch file to start VAC
This is a simple check for any currently running instances, then will start 2 instances, which is all I personally need. Then once those are running, it will start a simple monitor to check every X seconds to see if they are running or not. If not, then it will run this file. See __monitor_vac.bat for that.